swapchain

+++ crates/foxy_renderer/src/renderer/swapchain.rs
@@ -1,5 +1,6 @@
 use std::sync::Arc;
 
+use ezwin::window::Window;
 use itertools::Itertools;
 use tracing::*;
 use vulkano::{
@@ -10,7 +11,6 @@ use vulkano::{
   pipeline::graphics::viewport::Viewport,
   swapchain::{acquire_next_image, ColorSpace, Surface, Swapchain, SwapchainAcquireFuture, SwapchainCreateInfo},
 };
-use winit::window::Window;
 
 use crate::{error::RendererError, renderer::image_format::PresentMode};
 
@@ -39,7 +39,7 @@ impl FoxySwapchain {
   ) -> Result<Self, RendererError> {
     let window_size = window.inner_size();
     let extent_2d = window.inner_size().into();
-    let extent_3d = [window_size.width, window_size.height, 1];
+    let extent_3d = [window_size.width as u32, window_size.height as u32, 1];
     let mut viewport = Viewport {
       offset: [0.0, 0.0],
       extent: [0.0, 0.0],
